Moved migrations to separate file.
