+++ rubbergod.py
@@ -1,5 +1,5 @@
-import re
 import traceback
+import argparse
 
 from discord.ext import commands
 
@@ -13,76 +13,28 @@ from repository.database.verification import Permit, Valid_person
 from repository.database.image import Image
 from repository.review_repo import ReviewRepository
 
+import repository.db_migrations as migrations
 
-# TODO move this ANYWHERE else
-def load_dump():
-    database.base.metadata.create_all(database.db)
+parser = argparse.ArgumentParser()
+parser.add_argument('--load_dump', type=str,
+                    help='Imports SQL dump from SQL file to database.',
+                    metavar='filepath.sql')
+parser.add_argument('--load_subjects', action='store_true',
+                    help='Fills DB with subjects.')
+parser.add_argument('--init_db', action='store_true',
+                    help='Creates missing DB tables.')
+args = parser.parse_args()
 
-    session.query(Karma).delete()
-    session.query(Karma_emoji).delete()
-    session.query(Permit).delete()
-    session.query(Valid_person).delete()
-    session.commit()
-
-    data = database.base.metadata.tables.keys()
-    for row in data:
-        print(row)
-
-    with open("backup-2019-08-17.sql", "r") as backup_file:
-        data = backup_file.readlines()
-
-    inserts = [line for line in data if line.startswith("INSERT")]
-
-    karma_values = []
-    for insert in inserts:
-        values = insert.split("VALUES", 1)[1]
-        if insert.startswith("INSERT INTO `bot_karma`"):
-            values = values[1:-2].replace('\'', '')
-            values = values.replace('(', '').replace(')', '')
-            values = values.split(',')
-            for i in range(0, len(values), 3):
-                karma_values.append(Karma(member_ID=values[i],
-                                          karma=values[i + 1]))
-        elif insert.startswith("INSERT INTO `bot_karma_giving`"):
-            values = values[1:-2].replace('\'', '')
-            values = values.replace('(', '').replace(')', '')
-            values = values.split(',')
-            for i in range(0, len(values), 4):
-                karma_values.append(Karma(member_ID=values[i],
-                                          positive=values[i + 1],
-                                          negative=values[i + 2]))
-        elif insert.startswith("INSERT INTO `bot_karma_emoji`"):
-            values = values[1:-2].replace('\'', '')
-            values = values.replace('(', '').replace(')', '')
-            values = values.split(',')
-            for i in range(0, len(values), 2):
-                session.add(Karma_emoji(emoji_ID=values[i],
-                                        value=values[i + 1]))
-        elif insert.startswith("INSERT INTO `bot_permit`"):
-            values = values[1:-2]
-            values = values.replace('(', '').replace(')', '')
-            values = re.split(r',(?=\')', values)
-            values = [value.replace('\'', '') for value in values]
-            for i in range(0, len(values), 3):
-                session.add(Permit(login=values[i],
-                                   discord_ID=values[i + 2]))
-        elif insert.startswith("INSERT INTO `bot_valid_persons`"):
-            values = values[1:-2].replace('\'', '')
-            values = values.replace('(', '').replace(')', '')
-            values = values.split(',')
-            for i in range(0, len(values), 5):
-                session.add(Valid_person(login=values[i],
-                                         name=values[i + 1],
-                                         year=values[i + 2],
-                                         code=values[i + 3]
-                                         if values[i + 3] != "NULL" else None,
-                                         status=values[i + 4]))
-
-    for karma in karma_values:
-        session.merge(karma)
-
-    session.commit()
-
+if args.load_dump is not None:
+    migrations.load_dump(args.load_dump)
+    exit(0)
+elif args.load_subjects:
+    migrations.load_subjects()
+    exit(0)
+elif args.init_db:
+    migrations.init_db()
+    print('Init complete')
+    exit(0)
 
 config = config.Config
 
@@ -92,17 +44,6 @@ bot = commands.Bot(
                    case_insensitive=True)
 
 presence = presence.Presence(bot)
-
-
-def load_subjects():
-    """
-    Fills DB with subject shorcut from config file.
-    This is needed for reviews feature.
-    Run this just when you want to create DB fo reviews.
-    """
-    review_repo = ReviewRepository()
-    for subject in config.subjects:
-        review_repo.add_subject(subject)
 
 
 @bot.event
@@ -170,11 +111,6 @@ async def missing_arg_error(ctx, error):
         await ctx.send('Missing argument.')
     if isinstance(error, commands.errors.CheckFailure):
         await ctx.send(utils.fill_message("insufficient_rights", user=ctx.author.id))
-
-database.base.metadata.create_all(database.db)
-session.commit()  # Making sure
-
-# load_subjects()
 
 for extension in config.extensions:
     bot.load_extension(f'cogs.{extension}')
